Add nested structures support
In _wrap_g_field_info_set_value were added a conditional with a few
cases that are not handled by g_field_info_set_field. So far,
only structures are handled.

Ditto in _wrap_g_field_info_get_value.

In the Python base class for structures, the counterpart of __eq__, the
__ne__ magic, was needed too to ensure coherent comparisons.