Rewrite the invocation checks in C
Callable.type_check has been implemented in C in
pyg_argument_from_pyobject_check, which is called by
_wrap_g_function_info_invoke for each input argument.

Voids are not checkable and booleans don't need checks because every
Python object has a truth value.

Most numeric types are first tested with Py(Int|Long|Float)_Check.

Integers are also tested against the maximum and minimum value (for both
Python integers and longs) to be sure they fit into the 8-bit, 16-bit,
etc. argument. Sometimes, it is simpler since we know that Python
integers are longs, or because the
Py(Int|Long)_As(Long|UnsignedLong|...) functions will do the job for us
and raise an error if it can't convert the value.

Python float and doubles are converted to doubles and their value is
checked as well.

Strings are only tested with PyString_Check.

All other type checks are mostly incomplete and need further work.