Get rid of gi.repository
The only interesting feature of the module was the modules cache, which can be
implemented directly in the importer.
The ref. count of the repository in _wrap_g_irepository_get_default must be
incremented before returning the object, since we keep a static reference.
I also offered _wrap_g_irepository_get_default a little style enhancement.