Use the right object when checking numbers
In _pygi_g_type_info_check_object, numbers were checked against the object
which had to implement the number protocol. Although such an object may be
correct, the check required the comparison functions to be implemented in that
object too.

The right way to check those objects is to check the converted
number (an int, long or float) for the bounds, not the original object.