Add capabilities to import wrappers from pygi
At instance creation for boxed and pointers, at lookup for objects,
when the gtype has no wrapper yet, a wrapper may be imported from pygi.

The feature is turned on at configure time by --enable-pygi.

Because we couldn't create a circular build dependency, PyGI's import function and
API definition had to be copied in this tree.