Create overridden modules in two passes
This patch splits overridden module creation into two passes. The first pass
creates the auto-generated module normally before the overridden module is
attempted to be imported. The second pass imports the overridden module and
replaces the auto-generated module with the overridden. This is necessary
for the overridden modules to be able to access the auto-generated ones.