Fix members initialization in metaclasses
In metaclasses, the test for the name of the class was wrong, since it
prevented one to create a subclass with the same name (especially annoying for
overrides). Now, if a GType is available from the info, the fact that it
doesn't have any wrapper yet means that the metaclass is creating the base
class, which will be registerd just after its creation. This is true for
objects, and for structures registered as boxed or pointer too.

This patch includes a test for basic subclassing in Python. It notably tests
that methods don't get overridden by the metaclass.