Restore the overrides support
Add a ModuleProxy in front of the DynamicModule when an overrides module is
present. There is no need for an overrides module to be a class; it can just be a module.

Add an override decorator to override the wrapper of a registered type.

Adapt Gdk and Gtk accordingly.

Add tests.

https://bugzilla.gnome.org/show_bug.cgi?id=602830