Force out arguments to be initialized as NULL.  Comes with a test.
This fix was motivated by a real world library which had a transfer full
utf8 out argument which sometimes was not set.  We would leave the pointer
dangling and try and free it at the end of invoke() and crash.  Library refused
to change their behavior so we're forced to take care of it on our end.

https://bugzilla.gnome.org/show_bug.cgi?id=616043