tests: Port to new introspection tests
Everything is renamed "Regress", and both it and GIMarshallingTests
are now in source form, so we compile them.

The scanner now adds "static methods" to objects, structs, and unions,
so update the test code to use those.

In the tests, remove broken (inout) cases - the person writing these
tests misunderstood the semantics of (inout).  It's not acceptable for
a C API to mutate e.g. a GSList* passed in, or unref an object.

The invocation code needed to be updated for this - remove some
broken hacks.

https://bugzilla.gnome.org/show_bug.cgi?id=627878