pygi: always free the invocation_state struct
In pygi-invoke.c, the invocation_state struct is never freed
in case of success. Thus, always call _free_invocation_state()
before leaving.
Modify _free_invocation_state to avoid double free in case of
caller-allocated GValue, once as a released argument in the
_process routine and another time in the _free as the special
case. So move all argument releasing code from the _process
routine to the _free one.
Modify the tests for the callback routines to return an integer
value as specified in the GIR file.

Make check is as successful as before (already existing error
related to GVariant is still there).

https://bugzilla.gnome.org/show_bug.cgi?id=630271