Move pyglib_{main_context, option_context, option_group}_new into _PyGLib_API
_PyG{MainContext, OptionContext, and OptionGroup_Type} were not be initialized
when used inside the glib module, since pyglib_init is not called.

pyglib.c is compiled as a stand-alone library loaded by the _glib module that
declares the above-mentioned types. Hence, they cannot be accessed by the
former. This patch moves the functions that need those symbols into the
glib._glib module and exports them to the pyglib library through _PyGLib_API.

https://bugzilla.gnome.org/show_bug.cgi?id=636656