Fix the __dir__() methods on DynamicModule and IntrospectionModule
Previously the __dir__() methods did not list all attributes.
A simple test case is included. It does not test to see if
every attribute is listed, it just tests a few of each kind:
- (wrapped) typelib attributes
- class attributes and methods
- instance attributes

A set() is used to avoid returning duplicate attributes.
The test case checks for this as well.

https://bugzilla.gnome.org/show_bug.cgi?id=639229