In IntrospectionModule and DynamicModule classes, make all instance attributes start with an underscore.
This changes IntrospectionModule.version to _version and DynamicModule.introspection_module to _introspection_module.
This is done to mark the attributes as private, and also avoid name collisions with attributes from the typelib.
In Gstreamer, there is a function gst_version, which was previously inaccessible because of IntrospectionModule.version overriding it.