Always register a new GType when a GObject class is subclassed
This patch makes the GType <-> python mapping much more predictible,
and fixes the bug caused by overriding methods without specifying a
__gtype_name__ member in the subclass, and makes type_register useless
for real :-)

It is still possible to provide an explicit __gtype_name__ member in the
subclass as it allows having a predictible GType name, which is handy
for some of our tests. There is also an explicit special case for
overrides because we obviously do not want to register new GTypes for
those ones as it would clearly defeat the purpose of overrides.

https://bugzilla.gnome.org/show_bug.cgi?id=543056