Modify override for Gtk.Adjustment to allow position or keyword arguments in __init__().
Previously passing no arguments was not working, because the default value for each parameter was None, and GObject.__init__() refuses to allow None for integer properties. This patch does not pass None up to GObject.__init__. Instead it does not pass the parameter at all, and uses the class's default values.

https://bugzilla.gnome.org/show_bug.cgi?id=639934