Revert "Fix Pango FontDescription override"
According to
http://library.gnome.org/devel/pango/1.28/pango-Fonts.html#pango-font-description-new
the default constructor actually does take no arguments; we should actually fix
the MemoryError. Add a test case for this.

Remove the FIXME though, as pango_font_description_from_string() is not a
FontDescription constructor, but a static factory method.

Thanks to Paolo Borelli for pointing this out!

This reverts commit 8878c57676091c08e66bc6cbe735d898cb420582.