Speed up class creation: rewrite _setup_vfuncs() to be much more efficient.
This patch rewrites the _setup_vfuncs() method to remove recursion and
make the running time linear in the number of virtual functions to hook up
(ie. methods starting with "do_") instead of linear in the number of
virtual functions in the base class which could possibly be overridden.

Since most classes do not override all of the virtual functions in the
base class (and many override none), this runs much faster.

It is possible to not recurse on all base classes because non-interface
base classes will have the virtual function installed as an attribute.
Thus getattr() can be called, which recurses to the base classes much
faster than a custom implementation in Python. If the method cannot be
found with getattr(), all interface bases classes are searched manually.

The function is_function_in_classes() has been deleted. Because of the
above changes, it is not used anymore.

https://bugzilla.gnome.org/show_bug.cgi?id=640073