Speed up _setup_native_vfuncs()
This changes _setup_native_vfuncs() to only install native
vfunc wrappers from the current class on the current class.
Native vfuncs will not be propogated up or down the class
hierarchy as this is unnecessary and wastes CPU and memory.

Since the normal process in python to retrieve a method or
attribute recurses to the base classes if an attribute is not
found in the subclass, there is no need to setup all base class
virtual functions on a subclass.

This patch removes the recursion in _setup_native_vfuncs()
and lets Python find them in the base classes like a normal
Python class would work. This significantly increases the speed
of any class which is or inherits from a C class which includes
virtual methods.

https://bugzilla.gnome.org/show_bug.cgi?id=640629