[gi] Accept only a single object in GLib.Variant constructor
We previously allowed flat arguments for tuple signatures, e. g.

  GLib.Variant('(ii)', 1, 2)

However, that's not how GVariant is supposed to work. Remove the special case
to handle flat argument lists, and only accept a single value, i. e.

  GLib.Variant('(ii)', (1, 2))

Note that this breaks the current API, but as it is not used widely yet, let's
better fix it now.

Thanks to Ryan Lortie for pointing this out!