[gi] Add Pythonic gdbus method invocation
Provide a wrapper for Gio.DBusProxy for calling D-Bus methods like on a normal
Python object. This will handle the Python object <-> GVariant conversion, and
optional keyword arguments for flags, timeout, and a result handler for
asynchronous calls.

Require specifying the input argument signature as the first argument of each
method call. This ensures that the types of e. g. integers are always correct,
and avoids having to do expensive D-Bus introspection for each call.

https://bugzilla.gnome.org/show_bug.cgi?id=640181