[tests] Separate processes for GI and static binding tests.
Importing and using both static gobject bindings and
introspection GObject bindings in the same process can cause
conflicts with types which otherwise wouldn't be there.

This patch changes "make check" to call runtests.py twice -- once
for each set of tests.

In the case of a test failure, runtests.py now sets the exit code
so that make does not continue. Otherwise you might miss the
failures from the first batch of tests in the scrollback.