Fix flags with multiple names for the same value.
Flags constructs a dict __flags_values__ and uses it to cache
instances. However some flags in Glib such as G_IO_FLAG_MASK and
G_IO_FLAG_GET_MASK are aliases for the same int value, and will
override each other's place in the dictionary.

The dict length check is not necessary. It only reduces the number
of duplicate instances we keep, because if an instance is not
found in the dict, a new one is created anyway.