Fix ABI break in old static bindings.
Commit 84d6142c14a7ebfb7284d3db52e14d3393f93905 (Always register a new
GType when a GObject class is subclassed) breaks the more advanced usage
of PyGObject with regards to "metaclass hackery" as used in for example
the kiwi and sqlkit projects. But the users of the gi-based bindings
now rely on the new behaviour.

We fix this by restraining the systematical registering of new types to
the new gi-based bindings, leaving the old pygtk ones untouched.

https://bugzilla.gnome.org/show_bug.cgi?id=646437