pygmainloop: avoid lockups if multiple glib.MainLoop exist
If multiple glib.MainLoop() sources exist, then we will add multiple
watches, and when python writes a byte to the wakeup pipe, all of the
sources will try to read it; only one will succeed and the others
will block. Set both ends of the pipe nonblocking to avoid this.

https://bugzilla.gnome.org/show_bug.cgi?id=663068