Fix a few set-but-not-used warnings.
In a couple of cases, the variable in question was set to a value
spelled out again later in the function.

The 'sequence_cache' variable is re-declared five lines below.

The return value of 'read' was previously completely ignored. The
'gssize ret' variable was in fact added to squash an unused-result
warning.

https://bugzilla.gnome.org/show_bug.cgi?id=668497