Atomic inserts in Gtk.{List,Tree}Store overrides
Gtk.{List,Tree}Store's overrides provide append(), insert() etc. methods which
take an optional data row array. If this is given, use insert_with_valuesv()
instead of creating a new iter and then filling it with data. The latter sent a
row-added signal, at which time the row was still empty, and a subsequent
row-changed signal. With this we only get a single row-added signal with
complete row data.

Note that this does not change insert_{before,after}(), as there is no
counterpart of insert_with_valuesv() which takes a TreeIter instead of a
position. For those you will still get two signals, and have to deal with None
values.

https://bugzilla.gnome.org/show_bug.cgi?id=671610