Correct the Gtk.Window hierarchy
We need to make sure that all Gtk.Dialog subclasses inherit from
the overridden Window class. For that to be done automaticly we need
to create the Window class before the Dialog class.

Now when it's inherited properly we need to avoid calling the Window
constructor twice as it passes in a construct-only parameter. So add
**kwargs to the Window constructor to allow us to pass in any kind
of GObject property to it and refactor the Dialog subclasses to pass
in all properties to the same constructor.

Also adds a bunch of tests to make sure that the hiearchy is correct.

https://bugzilla.gnome.org/show_bug.cgi?id=672158

Signed-off-by: Martin Pitt <martin.pitt@ubuntu.com>