Make tests fail on CRITICAL logs, too, and apply to all tests
Instead of setting warnings/criticals to fatal in individual test modules, do
it in runtests.py, so that it applies to all tests.

We currently have some tests which are known to generate CRITICALs (now marked
with FIXME), and some WARNINGs (as they test behaviour with known-bad values).
For these, warnings/criticals are now explicitly permitted.