Add special case for Gdk.Atom array entries from Python
Gdk.Atom pretends to be a struct pointer, but is really just an int wrapped
into a pointer. So we must not dereference it directly, nor free it, but
instead just copy the pointer value.

Also add a few other test cases for "single Atom return", "single Atom argument
in", and Atom GList return", which already work fine.

https://bugzilla.gnome.org/show_bug.cgi?id=661709