[API add] Add GObject.bind_property method
This adds the "bind_property" method for binding two gobject properties
together. The method returns a weak reference to a GBinding object.
The BindingWeakRef object is used to manage GBinding objects within python
created through GObject.bind_property. It is a sub-class PyGObjectWeakRef so
that we can maintain the same reference counting semantics between Python
and GObject Binding objects. This gives explicit direct control of the
binding lifetime by using the "unbind" method on the BindingWeakRef object
along with implicit management based on the lifetime of the source or
target objects.

Note this does not yet include support for converter closures. This can come
later after the initial implementation is accepted.

https://bugzilla.gnome.org/show_bug.cgi?id=675582

Signed-off-by: Martin Pitt <martinpitt@gnome.org>