Escape identifiers which are Python keywords
Add a trailing underscore to identifiers which are Python keywords.

We use a per-major-version static identifier list derived from keyword.kwlist
instead of calling out to Python's keyword.iskeyword(). This is much faster,
and also allows us to tweak the result. For example, Python 3 dropped "print"
as a keyword, but we still want to escape that to avoid breaking the API
between different Python versions.

Error out when building with a major Python version not covered yet, so that we
do not forget to update the list in the future.

https://bugzilla.gnome.org/show_bug.cgi?id=676746