Remove "label" property from Gtk.MenuItem if it is not set
The Gtk.MenuItem will not render as a separator if the "label" or
"user-underline" properties have been accessed. The constructor
for Gtk.MenuItem override should not pass the "label" property
as an argument if it is None since that will still result in an
empty label widget which breaks Gtk.SeparatorMenuItem.

https://bugzilla.gnome.org/show_bug.cgi?id=670575

Signed-off-by: Martin Pitt <martinpitt@gnome.org>