Fix array handling for interfaces, properties, and signals
Fix lots of corner cases where arrays are not handled properly.
_pygi_argument_to_object() now has the documented expectation of getting arrays
packed in GArrays. This was implicit before and not correctly done on most call
sites.

The helper _pygi_argument_to_array() has been improved to work on any kind of
array. Fix all call sites of _pygi_argument_to_object() to do the
array conversion appropriately before calling _pygi_argument_to_object().

Adds a test case that implements a GInterface with a method that takes an array
of variants as input.

https://bugzilla.gnome.org/show_bug.cgi?id=667244