Fix error messages on interface/class type mismatches
Previously, when you called a function with an argument which was not
compatible with the expected class/interface type, you got an error message
like

  TypeError: Expected Gtk.TreeViewColumn, but got GObjectMeta

which had the wrong (and useless) class name for the actual type, and did not
tell you which argument caused the problem. With this it says e. g.

  TypeError: argument column: Expected Gtk.TreeViewColumn, but got Gtk.Button

instead.