Fix list marshalling on big-endian machines
On big endian machines we cannot simply set e. g. GIArgument.v_int8 and expect
GIArgument.v_pointer to be a correct representation. This needs to use
GINT_TO_POINTER/GPOINTER_TO_INT properly, so use the already existing
_pygi_hash_pointer_to_arg()/_pygi_arg_to_hash_pointer() methods in marshalling
to and from GList and GSList, and handle int8 and int16 as well.

Part of porting pygobject to ppc64:
https://bugzilla.redhat.com/show_bug.cgi?id=842880
https://bugzilla.gnome.org/show_bug.cgi?id=680693