Fix erronous import statements for Python 3.3
Update pygobject-external.h to use GType (which is what GTypeWrapper is
exposed as) instead of GTypeWrapper when attempting import.

Catch ImportError around attempted imports of a typelibs override file
which don't always exist (GObject...). This is a behavioural change in
Python 3.3 (http://bugs.python.org/issue15715), but let's fix it anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=682051