Fix for python 2.6, drop support for < 2.6
Replace sys.version_info.major access to tuple access which also works for
Python 2.6.

When building for Python 2.6, inject some missing unittest API such as
@unittest.skipUnless and assertGreaterEqual() into the unittest module in
runtests.py, so that the tests have a chance to run.

As building with Python 2.5 has been broken for a long time with nobody
complaining, and 2.5 is ancient, bump minimum Python requirement to 2.6. Drop
obsolete #ifdef paths which only apply to <= 2.5.

https://bugzilla.gnome.org/show_bug.cgi?id=682422