pygtkcompat: Add pygtk compatible GenericTreeModel implementation
Add Python implementation of the GenericTreeModel that was
available in pygtk. The implementation attempts a better job
than the original at ref counting by guaranteeing no leaks
upon deletion of the model itself. Or by using the extra "node"
argument to the row_deleted signal. The model is available in
the pygtkcompat package directly as
pygtkcompat.generictreemodel.GenericTreeModel or with as
gtk.GenericTreeModel when pygtkcompat.enable_gtk() is set.

Add file list and tree demos making use of GenericTreeModel
to gtk-demo.

Auto-expand gtk-demo app tree to give a better overview of
the demos available.

https://bugzilla.gnome.org/show_bug.cgi?id=682933