Fix separate build tree and distcheck
Fix regression from c7c95a79: We must set sys.path in runtests.py properly and
cannot rely on setting $PYTHONPATH from Makefile.am only. Python always
prepends the directory of the source file to sys.path, but that points to the
source dir, not the build dir. The build dir has to take precedence, otherwise
we fail to import the built libraries.