Fix error messages for out of range numbers
PyErr_Format() does not understand %lli and %li, it needs to be %lld and %ld.
So we cannot use those and G_GINT64_FORMAT.

Also remove the "if (long_ < G_MININT64 || long_ > G_MAXINT64)" check, as long_
is a gint64 which can't possibly overflow its own data type. It would also have
an unprintable error message.

https://bugzilla.gnome.org/show_bug.cgi?id=684314