Change install_properties to not use getattr on classes
The usage of getattr for accessing a classes __gproperties__
variable can be problematic due to the potential of it returning
the parent classes variable when it does not exist on the sub-class.
Similar to the fix for https://bugzilla.gnome.org/show_bug.cgi?id=686496,
cls.__dict__.get is used to ensure this does not happen.

https://bugzilla.gnome.org/show_bug.cgi?id=686559