Fix TestSource.testSources() test case
PyGObject has established the assumption that the destruction of a GLib.Source
Python object does not destroy the actual GSource, as shown in
TestSource.setup_timeout(), TestTimeout.test504337(), and
https://bugzilla.gnome.org/show_bug.cgi?id=504337.

So we need to explicitly destroy our MySource and Idle objects after using
them, as their callbacks always return True and we do not want them to spill
over into other tests.

Also fix the assertions to actually verify that MySources' callback was callied
(pos > 0, not pos >= 0), and use the unittest comparison API instead of a
simple assert statement.

https://bugzilla.gnome.org/show_bug.cgi?id=686627