Add tests for priority argument of idle_add/timeout_add
There is a potential to treat the priority as user data in a call like
"GLib.idle_add(cb, GLib.PRIORITY_HIGH)". The current static bindings force
using a keyword argument for the priority (but silently ignore it if you
specify both userdata and priority as a positional argument).

Test the correct handling of priority as well.