Add tests for GLib.IOChannel
This did not have any code coverage at all. The tests uncover a range of bugs,
which will be fixed in the next commit.

This drops the verification of the set priority in
test_add_watch_data_priority(), as find_source_by_id() is not yet available in
this pygobject version.