[API change] Remove static filename_from_utf8() binding
Replace static GLib.filename_from_utf8() with GI. The old static binding always
tried to convert the result to an Unicode object, which will fail if the result
is not UTF-8 encoded (which is the whole point of this function really!), so
return bytes now.

Although the static binding was rather useless before, this is technically an
API break.