test_gi: Add missing data type tests from GIMarshallingTests
Add tests for types time_t, GBytes, GByteArray, and various variants of passing
around GValues, flags, enums, arrays, and boxed structs. This tests API which
is provided by GIMarshallingTests, but which we did not cover yet.

This detects a number of bugs in annotations and our marshalling. These have
been marked as expected failure, or disabled completely for the cases where
they cause crashes.