Fix array arguments on 32 bit
In _pygi_argument_from_object() we never put the actual element GType into the
constructed array. The array contains GIArguments, or bytes in the case of
passing a string as an array.

This happened to work on 64 bit machines where GIArgument and char* have the
same size, but not on 32 bit machines (test_gi.TestPropertiesObject.test_strv
fails there).