Fix test for GBytes.compare()
The result of the compare method is defined as equal, less than or greater than zero
and the test must match to that. The underlaying memcmp() function can return other
values than -1, 0 and 1. For example on architectures where it is implemented directly
via a CPU instruction like on s390(x) where I can see -2 as a result instead of the
"expected" -1.

https://bugzilla.gnome.org/show_bug.cgi?id=690837