Add __gpointer__ property to GObject static binding
Add access to the underlying C GObject pointer by wrapping it in a
PyCapsule/PyCPointer and exposing it as __gpointer__.
Add special case marshaling for gi parameters annotated as gpointer
to accept a PyCapsule and extract the underlying pointer as the arg.
This allows usage of methods like GObject.signal_handler_disconnect
which we can start replacing the static bindings with.

https://bugzilla.gnome.org/show_bug.cgi?id=692918