gi/pygi-marshal-from-py.c: Fix build on Visual C++
Since Visual C++ does not provide the INFINITY and NAN constants in its
math.h (they are items defined by C99), provide fallback implementations
for it here.  The INFINITY constant can be provided with HUGE_VAL, since
we are using INFINITY to check the value of a variable of double type, and
we could use a rather simple workaround for NAN.

Also avoid declaring variables in the middle of the block.

https://bugzilla.gnome.org/show_bug.cgi?id=692856