Fix marshalling of C arrays with explicit length in signal arguments
We need _pygi_argument_to_array() from both closure marshalling (where we have
the arguments as GIArgument array) and signal closure marshalling (where we
have the arguments in a GValue array). Add an alternative "args_values"
parameter to _pygi_argument_to_array() so that callers can specify one or the
other depending on which type they have available.

This allows us to pass on the full argument list for signal closures, so that
_pygi_argument_to_array() can access the explicit length argument for an
array.

This fixes the GSettings:change-event signal.

https://bugzilla.gnome.org/show_bug.cgi?id=662241