Deprecate GLib and GObject threads_init
Remove a handful of Python threading related helper functions
from pyglib and pygobject. The binding internals required
GLib.threads_init to be called for PyGObject to work with
Python threads. This was removed as it should not be a requirement.
Using the Python threading module already initializes threading
for us (PyEval_InitThreads).

https://bugzilla.gnome.org/show_bug.cgi?id=686914