Remove incorrect bounds check with property helper flags
Remove bounds check for flags as it is not necessary for the
helper to make these kind of judgement calls in general.
e.g. leave it to marshaling/internals to complain about potential
problems. The flags were being bounds checked to a maximum
value of 32 (the intention was most likely to limit it to 32 bits).