Optimize signal lookup in gi repository
Now that we have GSignalQuery results available to us when connecting
signals, we already know which GType implements the signal in question.

Therefore there is no need to traverse the class parent hierarchy
looking for this, which takes a considerable amount of CPU time.

There is also no need to canonicalize the signal name; both before
and after this patch, by the time we reach this point we have already
successfully looked up the signal name as presented from Python.

https://bugzilla.gnome.org/show_bug.cgi?id=696143