testhelper: Fix import requirement for GObject
Replace the importing of gi._gobject._gobject with gi.repository.GObject
in tests/testhelpermodule.c

The testhelper module was only importing the static bindings
(gi._gobject._gobject) and not the overrides (gi.repository.GObject).
This was causing some tests to fail when test_thread was the first
test to run in the suite due to it registering new types based on
PyGObject_Type.

https://bugzilla.gnome.org/show_bug.cgi?id=703647