Move to Python basic type marshaling out of _pygi_argument_to_object
Move the marshaling of GI arguments to Python objects for basic types
into a new function. The required information for this marshaler
is a GITypeTag and GITransfer. Argument marshaling matching these
requirments are now found in: _pygi_argument_to_object_basic_type.
The new marshaler can be used with a generic argument cache marshaler
to unify all of the "basic type" marshaling.

https://bugzilla.gnome.org/show_bug.cgi?id=693405