Cleanup per-item array marshaling code for flat arrays
Add an early per-item check which tests if the item being marshaled is a
pointer and simply copies the pointer into the array. This takes care of the
GdkAtom and GVariant special cases because these items are always reported
as pointers.
Fix error condition cleanup code when an item fails marshaling in the middle
of an array.

https://bugzilla.gnome.org/show_bug.cgi?id=693402