Fix TypeError when setting drag target_list to None
When calling Widget.drag_dest_set_target_list(None) or Widget.drag_source_set_target_list(None)
a "TypeError: 'NoneType' object is not iterable" is thrown.

According to Gtk documentation [1] [2], this shouldn't be the case since client code should be
able to pass NULL in calls made to gtk_drag_dest_set_target_list and
gtk_drag_source_set_target_list.

We now check if the target_list is None and do not try to create a TargetList if it is the case.

[1] https://developer.gnome.org/gtk3/3.10/gtk3-Drag-and-Drop.html#gtk-drag-dest-set-target-list
[2] https://developer.gnome.org/gtk3/3.10/gtk3-Drag-and-Drop.html#gtk-drag-source-set-target-list

https://bugzilla.gnome.org/show_bug.cgi?id=709926