Remove special case GObject base class check when creating GI classes
Replace explicit GObject.Object string name check when calculating the
introspection class hierarchy with a more generalized technique. This allows
any C based wrapper of a GType to "underride" an introspection class
automatically. This currently only handles the case of GObject.Object, but
will be used for fundamentals and GParamSpec.

https://bugzilla.gnome.org/show_bug.cgi?id=631901