Add threads_init back as a requirement for non-Python threaded repos
Re-add a "threads_init" function to gi for explicit intialization of Python
threading support. This was marked as deprecated in the previous cycle
because using Python threads already initializes everything.
However, we still need an explicit initalization when using repositories
with non-Python threads which may interact with Python callbacks
(GStreamer).

https://bugzilla.gnome.org/show_bug.cgi?id=710447