Remove overzealous argument checking for callback userdata
Remove check which ensures userdata is None if the callback is None.
This check would need to become more complicated with recent versions of
PyGObject where userdata can be variable (would also need to check against
a tuple containing None). Instead of adding more complex checking, simply
remove the checking as it is unnecessary to begin with.

https://bugzilla.gnome.org/show_bug.cgi?id=711173