Replace usage of PyGIBoxed_Type with PyGIStruct_Type
Remove empty definition of PyGIBoxed_Type and use PyGIStruct_Type in its
place for GI_INFO_TYPE_BOXED based infos in repositories. As noted in
the docs: "GI_INFO_TYPE_BOXED boxed, see GIStructInfo or GIUnionInfo"
we can use the GIStructInfo API for boxed types because we also dispatch
base classes on GType checks as seen in:
https://git.gnome.org/browse/pygobject/tree/gi/module.py?id=3.11.2#n186
This fixes some of the issues noted in bug 581525.

https://bugzilla.gnome.org/show_bug.cgi?id=581525